refactor : 설정클래스 config 패키지로 이동
