+++ src/main/kotlin/com/flab/inqueue/config/jpa/JpaAuditingConfiguration.kt
@@ -1,0 +1,8 @@
+package com.flab.inqueue.config.jpa
+
+import org.springframework.context.annotation.Configuration
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing
+
+@EnableJpaAuditing
+@Configuration
+class JpaAuditingConfiguration
